wifi: resolved issue with wifi_getchar

+++ cc_code/src/esp.h
@@ -9,7 +9,7 @@
 #define ESP_H_
 
 void setup_uart1();
-uint8_t wifi_sendchar(int txChar);
+uint8_t wifi_sendchar(uint8_t txChar);
 uint8_t wifi_getchar(void);
 char * wifi_sendstring(char * cmd);
 char wifi_checkstring(char * response);
